Move bqrs-cli-types.ts to /common
