+++ extensions/ql-vscode/src/data-extensions-editor/external-api-usage.ts
@@ -1,4 +1,4 @@
-import { ResolvableLocationValue } from "../pure/bqrs-cli-types";
+import { ResolvableLocationValue } from "../common/bqrs-cli-types";
 
 export type Call = {
   label: string;
