+++ extensions/ql-vscode/src/local-queries/results-view.ts
@@ -64,7 +64,7 @@ import {
   RawResultSet,
   transformBqrsResultSet,
   ResultSetSchema,
-} from "../pure/bqrs-cli-types";
+} from "../common/bqrs-cli-types";
 import {
   AbstractWebview,
   WebviewPanelConfig,
