+++ extensions/ql-vscode/src/run-queries-shared.ts
@@ -27,7 +27,7 @@ import { nanoid } from "nanoid";
 import { CodeQLCliServer } from "./codeql-cli/cli";
 import { SELECT_QUERY_NAME } from "./language-support";
 import { DatabaseManager } from "./databases/local-databases";
-import { DecodedBqrsChunk, EntityValue } from "./pure/bqrs-cli-types";
+import { DecodedBqrsChunk, EntityValue } from "./common/bqrs-cli-types";
 import { BaseLogger, extLogger } from "./common";
 import { generateSummarySymbolsFile } from "./log-insights/summary-parser";
 import { getErrorMessage } from "./pure/helpers-pure";
