+++ extensions/ql-vscode/src/view/compare/CompareTable.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { SetComparisonsMessage } from "../../pure/interface-types";
 import RawTableHeader from "../results/RawTableHeader";
 import { className } from "../results/result-table-utils";
-import { ResultRow } from "../../pure/bqrs-cli-types";
+import { ResultRow } from "../../common/bqrs-cli-types";
 import RawTableRow from "../results/RawTableRow";
 import { vscode } from "../vscode-api";
 import { sendTelemetry } from "../common/telemetry";
