+++ extensions/ql-vscode/src/view/results/RawTableHeader.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { vscode } from "../vscode-api";
 import { RawResultsSortState, SortDirection } from "../../pure/interface-types";
 import { nextSortDirection } from "./result-table-utils";
-import { Column } from "../../pure/bqrs-cli-types";
+import { Column } from "../../common/bqrs-cli-types";
 
 interface Props {
   readonly columns: readonly Column[];
