+++ extensions/ql-vscode/src/view/results/__tests__/results.spec.tsx
@@ -8,7 +8,7 @@ import {
 } from "../../../pure/interface-types";
 import * as fs from "fs-extra";
 import { resolve } from "path";
-import { ColumnKindCode } from "../../../pure/bqrs-cli-types";
+import { ColumnKindCode } from "../../../common/bqrs-cli-types";
 import { postMessage } from "../../common/post-message";
 
 const exampleSarif = fs.readJSONSync(
