+++ extensions/ql-vscode/test/vscode-tests/cli-integration/new-query.test.ts
@@ -4,7 +4,7 @@ import { CancellationTokenSource } from "vscode-jsonrpc";
 import * as messages from "../../../src/pure/new-messages";
 import * as qsClient from "../../../src/query-server/query-server-client";
 import * as cli from "../../../src/codeql-cli/cli";
-import { CellValue } from "../../../src/pure/bqrs-cli-types";
+import { CellValue } from "../../../src/common/bqrs-cli-types";
 import { describeWithCodeQL } from "../cli";
 import { QueryServerClient } from "../../../src/query-server/query-server-client";
 import { extLogger, ProgressReporter } from "../../../src/common";
