+++ extensions/ql-vscode/test/vscode-tests/no-workspace/run-queries.test.ts
@@ -22,7 +22,7 @@ import {
 } from "../../../src/query-server/legacy";
 import { DatabaseItem } from "../../../src/databases/local-databases";
 import { DeepPartial, mockedObject } from "../utils/mocking.helpers";
-import { BqrsKind } from "../../../src/pure/bqrs-cli-types";
+import { BqrsKind } from "../../../src/common/bqrs-cli-types";
 
 describe("run-queries", () => {
   let isCanarySpy: jest.SpiedFunction<typeof config.isCanary>;
